Enum assembly_fdb::core::Value

pub enum Value<T>
where
    T: Context,
{
    Nothing,
    Integer(i32),
    Float(f32),
    Text(<T as Context>::String),
    Boolean(bool),
    BigInt(<T as Context>::I64),
    VarChar(<T as Context>::XML),
}

A single field value in the database

This is a generic enum that is the template for all other Field types in this crate.

Variants

Nothing

The NULL value

Integer(i32)

A 32 bit integer

Float(f32)

A 32 bit IEEE floating point number

Text(<T as Context>::String)

A string

Boolean(bool)

A boolean

BigInt(<T as Context>::I64)

A 64 bit integer

VarChar(<T as Context>::XML)

A (XML?) string

Implementations

impl<T> Value<T>
where T: Context,

pub fn map<O, M>(&self, mapper: &mut M) -> Value<O>
where O: Context, M: ValueMapperMut<T, O>,

Creates a value of a different context using the given mapper

pub fn into_opt_integer(self) -> Option<i32>

Returns Some with the value if the field contains an Value::Integer.

pub fn into_opt_float(self) -> Option<f32>

Returns Some with the value if the field contains a Value::Float.

pub fn into_opt_text(self) -> Option<<T as Context>::String>

Returns Some with the value if the field contains a Value::Text.

pub fn into_opt_boolean(self) -> Option<bool>

Returns Some with the value if the field contains a Value::Boolean.

pub fn into_opt_big_int(self) -> Option<<T as Context>::I64>

Returns Some with the value if the field contains a Value::BigInt.

pub fn into_opt_varchar(self) -> Option<<T as Context>::XML>

Returns Some with the value if the field contains a Value::VarChar.