Module assembly_fdb::core::loader

General-Purpose file loader

This is the original entry point to the FDB loading API. A SchemaLoader wraps an implementation of BufRead and loads the data from the file into an instance of Schema.

This uses the methods defined in the reader module and produces the data structure defined in the core module.

Structs

LoaderConfigImpl

Configuration for SchemaLoader

SchemaLoader

Structure to load a schema from some encapsulated stream

Traits

LoaderConfig

Configuration for the SchemaLoader