Struct assembly_fdb::mem::Table

#[repr(C)]
pub struct Table<'a> { /* private fields */ }

Reference to a single table

Implementations

impl<'a> Table<'a>

pub fn name_raw(&self) -> &'a Latin1Str

Get the undecoded name of the table

pub fn name(&self) -> Cow<'a, str>

Get the name of the table

pub fn index_iter(&self, id: u32) -> impl Iterator<Item = Row<'a>>

Get a list of rows by index

pub fn bucket_index_iter(&self, id: u32) -> TableRowIter<'a>

Get a list of all rows in the bucket of a given index

pub fn column_at(&self, index: usize) -> Option<Column<'a>>

Get the column at the index

Note: This does some computation, call only once per colum if possible

pub fn column_iter(&self) -> impl Iterator<Item = Column<'a>> + Clone

Get the column iterator

Note: This does some computation, call only once if possible

pub fn column_count(&self) -> usize

The amount of columns in this table

pub fn bucket_at(&self, index: usize) -> Option<Bucket<'a>>

Get the bucket at the index

Note: This does some computation, call only once per bucket if possible

pub fn bucket_for_hash(&self, hash: u32) -> Bucket<'a>

Get the bucket for the given hash

Note: This always calls Table::bucket_at exactly once

pub fn bucket_iter(&self) -> BucketIter<'a>

Get the bucket iterator

Note: This does some computation, call only once if possible

pub fn bucket_count(&self) -> usize

Get the amount of buckets

pub fn row_iter(&self) -> TableRowIter<'a>

Get an iterator over all rows