Trait assembly_fdb::parser::ParseLE

pub trait ParseLE: Sized + Copy {
    type Buf: AsMut<[u8]> + Default;

    const BYTE_COUNT: usize;

    fn parse(i: &[u8]) -> IResult<&[u8], Self>;
}

Marker trait that implies that Self can be parsed in little-endian mode

Associated Types

type Buf: AsMut<[u8]> + Default

A byte array of the same length that can be parsed as Self

Associated Constants

const BYTE_COUNT: usize

Same as std::mem::size_of::<Self>()

Required methods

fn parse(i: &[u8]) -> IResult<&[u8], Self>

Function to parse the buffer into self