Trait assembly_fdb::ro::buffer::BufferExt

pub trait BufferExt: Buffer {
    // Required methods
    fn get_len_at(&self, start: usize, len: usize) -> Res<&[u8]>;
    fn string(&self, addr: u32) -> Res<&Latin1Str>;
    fn i64(&self, addr: u32) -> Res<i64>;
    fn table_def_header(&self, addr: u32) -> Res<FDBTableDefHeader>;
    fn table_data_header(&self, addr: u32) -> Res<FDBTableDataHeader>;
    fn row_header_list_entry(&self, addr: u32) -> Res<FDBRowHeaderListEntry>;
    fn row_header(&self, addr: u32) -> Res<FDBRowHeader>;
}

Additional methods on &[u8]

Required Methods

fn get_len_at(&self, start: usize, len: usize) -> Res<&[u8]>

Get a subslice a the given offset of the given length

fn string(&self, addr: u32) -> Res<&Latin1Str>

Get a buffer as a latin1 string

fn i64(&self, addr: u32) -> Res<i64>

Get i64

fn table_def_header(&self, addr: u32) -> Res<FDBTableDefHeader>

Get the table definition header at the given addr.

fn table_data_header(&self, addr: u32) -> Res<FDBTableDataHeader>

Get the table data header at the given addr.

fn row_header_list_entry(&self, addr: u32) -> Res<FDBRowHeaderListEntry>

Get the FDBRowHeader list entry at the given addr.

fn row_header(&self, addr: u32) -> Res<FDBRowHeader>

Get the FDBRowHeader at the given addr.