Struct assembly_fdb::sqlite::Connection

pub struct Connection { /* private fields */ }

A connection to a SQLite database.

Implementations

impl Connection

pub fn busy_timeout(&self, timeout: Duration) -> Result<(), Error>

Set a busy handler that sleeps for a specified amount of time when a table is locked. The handler will sleep multiple times until at least “ms” milliseconds of sleeping have accumulated.

Calling this routine with an argument equal to zero turns off all busy handlers.

There can only be a single busy handler for a particular database connection at any given moment. If another busy handler was defined (using busy_handler) prior to calling this routine, that other busy handler is cleared.

Newly created connections currently have a default busy timeout of 5000ms, but this may be subject to change.

pub fn busy_handler( &self, callback: Option<fn(_: i32) -> bool>, ) -> Result<(), Error>

Register a callback to handle SQLITE_BUSY errors.

If the busy callback is None, then SQLITE_BUSY is returned immediately upon encountering the lock. The argument to the busy handler callback is the number of times that the busy handler has been invoked previously for the same locking event. If the busy callback returns false, then no additional attempts are made to access the database and SQLITE_BUSY is returned to the application. If the callback returns true, then another attempt is made to access the database and the cycle repeats.

There can only be a single busy handler defined for each database connection. Setting a new busy handler clears any previously set handler. Note that calling busy_timeout() or evaluating PRAGMA busy_timeout=N will change the busy handler and thus clear any previously set busy handler.

Newly created connections default to a busy_timeout() handler with a timeout of 5000ms, although this is subject to change.

impl Connection

pub fn prepare_cached(&self, sql: &str) -> Result<CachedStatement<'_>, Error>

Prepare a SQL statement for execution, returning a previously prepared (but not currently in-use) statement if one is available. The returned statement will be cached for reuse by future calls to prepare_cached once it is dropped.

fn insert_new_people(conn: &Connection) -> Result<()> {
    {
        let mut stmt = conn.prepare_cached("INSERT INTO People (name) VALUES (?)")?;
        stmt.execute(["Joe Smith"])?;
    }
    {
        // This will return the same underlying SQLite statement handle without
        // having to prepare it again.
        let mut stmt = conn.prepare_cached("INSERT INTO People (name) VALUES (?)")?;
        stmt.execute(["Bob Jones"])?;
    }
    Ok(())
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn set_prepared_statement_cache_capacity(&self, capacity: usize)

Set the maximum number of cached prepared statements this connection will hold. By default, a connection will hold a relatively small number of cached statements. If you need more, or know that you will not use cached statements, you can set the capacity manually using this method.

pub fn flush_prepared_statement_cache(&self)

Remove/finalize all prepared statements currently in the cache.

impl Connection

pub fn db_config(&self, config: DbConfig) -> Result<bool, Error>

Returns the current value of a config.

SQLITE_DBCONFIG_ENABLE_FKEY: return false or true to indicate whether FK enforcement is off or on

SQLITE_DBCONFIG_ENABLE_TRIGGER: return false or true to indicate whether triggers are disabled or enabled

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER: return false or true to indicate whether fts3_tokenizer are disabled or enabled

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE: return false to indicate checkpoints-on-close are not disabled or true if they are

SQLITE_DBCONFIG_ENABLE_QPSG: return false or true to indicate whether the QPSG is disabled or enabled

SQLITE_DBCONFIG_TRIGGER_EQP: return false to indicate output-for-trigger are not disabled or true if it is

pub fn set_db_config( &self, config: DbConfig, new_val: bool, ) -> Result<bool, Error>

Make configuration changes to a database connection

SQLITE_DBCONFIG_ENABLE_FKEY: false to disable FK enforcement, true to enable FK enforcement

SQLITE_DBCONFIG_ENABLE_TRIGGER: false to disable triggers, true to enable triggers

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER: false to disable fts3_tokenizer(), true to enable fts3_tokenizer()

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE: false (the default) to enable checkpoints-on-close, true to disable them

SQLITE_DBCONFIG_ENABLE_QPSG: false to disable the QPSG, true to enable QPSG

SQLITE_DBCONFIG_TRIGGER_EQP: false to disable output for trigger programs, true to enable it

impl Connection

pub fn pragma_query_value<T, F>( &self, schema_name: Option<DatabaseName<'_>>, pragma_name: &str, f: F, ) -> Result<T, Error>
where F: FnOnce(&Row<'_>) -> Result<T, Error>,

Query the current value of pragma_name.

Some pragmas will return multiple rows/values which cannot be retrieved with this method.

Prefer PRAGMA function introduced in SQLite 3.20: SELECT user_version FROM pragma_user_version;

pub fn pragma_query<F>( &self, schema_name: Option<DatabaseName<'_>>, pragma_name: &str, f: F, ) -> Result<(), Error>
where F: FnMut(&Row<'_>) -> Result<(), Error>,

Query the current rows/values of pragma_name.

Prefer PRAGMA function introduced in SQLite 3.20: SELECT * FROM pragma_collation_list;

pub fn pragma<F, V>( &self, schema_name: Option<DatabaseName<'_>>, pragma_name: &str, pragma_value: V, f: F, ) -> Result<(), Error>
where F: FnMut(&Row<'_>) -> Result<(), Error>, V: ToSql,

Query the current value(s) of pragma_name associated to pragma_value.

This method can be used with query-only pragmas which need an argument (e.g. table_info('one_tbl')) or pragmas which returns value(s) (e.g. integrity_check).

Prefer PRAGMA function introduced in SQLite 3.20: SELECT * FROM pragma_table_info(?);

pub fn pragma_update<V>( &self, schema_name: Option<DatabaseName<'_>>, pragma_name: &str, pragma_value: V, ) -> Result<(), Error>
where V: ToSql,

Set a new value to pragma_name.

Some pragmas will return the updated value which cannot be retrieved with this method.

pub fn pragma_update_and_check<F, T, V>( &self, schema_name: Option<DatabaseName<'_>>, pragma_name: &str, pragma_value: V, f: F, ) -> Result<T, Error>
where F: FnOnce(&Row<'_>) -> Result<T, Error>, V: ToSql,

Set a new value to pragma_name and return the updated value.

Only few pragmas automatically return the updated value.

impl Connection

pub fn transaction(&mut self) -> Result<Transaction<'_>, Error>

Begin a new transaction with the default behavior (DEFERRED).

The transaction defaults to rolling back when it is dropped. If you want the transaction to commit, you must call commit or set_drop_behavior(DropBehavior: :Commit).

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let tx = conn.transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

Failure

Will return Err if the underlying SQLite call fails.

pub fn transaction_with_behavior( &mut self, behavior: TransactionBehavior, ) -> Result<Transaction<'_>, Error>

Begin a new transaction with a specified behavior.

See transaction.

Failure

Will return Err if the underlying SQLite call fails.

pub fn unchecked_transaction(&self) -> Result<Transaction<'_>, Error>

Begin a new transaction with the default behavior (DEFERRED).

Attempt to open a nested transaction will result in a SQLite error. Connection::transaction prevents this at compile time by taking &mut self, but Connection::unchecked_transaction() may be used to defer the checking until runtime.

See Connection::transaction and [Transaction::new_unchecked] (which can be used if the default transaction behavior is undesirable).

Example

fn perform_queries(conn: Rc<Connection>) -> Result<()> {
    let tx = conn.unchecked_transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

Failure

Will return Err if the underlying SQLite call fails. The specific error returned if transactions are nested is currently unspecified.

pub fn savepoint(&mut self) -> Result<Savepoint<'_>, Error>

Begin a new savepoint with the default behavior (DEFERRED).

The savepoint defaults to rolling back when it is dropped. If you want the savepoint to commit, you must call commit or [set_drop_behavior(DropBehavior::Commit)](Savepoint:: set_drop_behavior).

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let sp = conn.savepoint()?;

    do_queries_part_1(&sp)?; // sp causes rollback if this fails
    do_queries_part_2(&sp)?; // sp causes rollback if this fails

    sp.commit()
}

Failure

Will return Err if the underlying SQLite call fails.

pub fn savepoint_with_name<T>( &mut self, name: T, ) -> Result<Savepoint<'_>, Error>
where T: Into<String>,

Begin a new savepoint with a specified name.

See savepoint.

Failure

Will return Err if the underlying SQLite call fails.

impl Connection

pub fn create_module<'vtab, T>( &self, module_name: &str, module: &'static Module<'vtab, T>, aux: Option<<T as VTab<'vtab>>::Aux>, ) -> Result<(), Error>
where T: VTab<'vtab>,

Register a virtual table implementation.

Step 3 of Creating New Virtual Table Implementations.

impl Connection

pub fn open<P>(path: P) -> Result<Connection, Error>
where P: AsRef<Path>,

Open a new connection to a SQLite database.

Connection::open(path) is equivalent to Connection::open_with_flags(path, OpenFlags::SQLITE_OPEN_READ_WRITE | OpenFlags::SQLITE_OPEN_CREATE).

fn open_my_db() -> Result<()> {
    let path = "./my_db.db3";
    let db = Connection::open(&path)?;
    println!("{}", db.is_autocommit());
    Ok(())
}

Failure

Will return Err if path cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn open_in_memory() -> Result<Connection, Error>

Open a new connection to an in-memory SQLite database.

Failure

Will return Err if the underlying SQLite open call fails.

pub fn open_with_flags<P>( path: P, flags: OpenFlags, ) -> Result<Connection, Error>
where P: AsRef<Path>,

Open a new connection to a SQLite database.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if path cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn open_with_flags_and_vfs<P>( path: P, flags: OpenFlags, vfs: &str, ) -> Result<Connection, Error>
where P: AsRef<Path>,

Open a new connection to a SQLite database using the specific flags and vfs name.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if either path or vfs cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn open_in_memory_with_flags(flags: OpenFlags) -> Result<Connection, Error>

Open a new connection to an in-memory SQLite database.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if the underlying SQLite open call fails.

pub fn open_in_memory_with_flags_and_vfs( flags: OpenFlags, vfs: &str, ) -> Result<Connection, Error>

Open a new connection to an in-memory SQLite database using the specific flags and vfs name.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if vfs` cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn execute_batch(&self, sql: &str) -> Result<(), Error>

Convenience method to run multiple SQL statements (that cannot take any parameters).

Example

fn create_tables(conn: &Connection) -> Result<()> {
    conn.execute_batch(
        "BEGIN;
         CREATE TABLE foo(x INTEGER);
         CREATE TABLE bar(y TEXT);
         COMMIT;",
    )
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn execute<P>(&self, sql: &str, params: P) -> Result<usize, Error>
where P: Params,

Convenience method to prepare and execute a single SQL statement.

On success, returns the number of rows that were changed or inserted or deleted (via sqlite3_changes).

Example

With positional params

fn update_rows(conn: &Connection) {
    match conn.execute("UPDATE foo SET bar = 'baz' WHERE qux = ?", [1i32]) {
        Ok(updated) => println!("{} rows were updated", updated),
        Err(err) => println!("update failed: {}", err),
    }
}

With positional params of varying types

fn update_rows(conn: &Connection) {
    match conn.execute(
        "UPDATE foo SET bar = 'baz' WHERE qux = ?1 AND quux = ?2",
        params![1i32, 1.5f64],
    ) {
        Ok(updated) => println!("{} rows were updated", updated),
        Err(err) => println!("update failed: {}", err),
    }
}

With named params

fn insert(conn: &Connection) -> Result<usize> {
    conn.execute(
        "INSERT INTO test (name) VALUES (:name)",
        &[(":name", "one")],
    )
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn path(&self) -> Option<&Path>

Returns the path to the database file, if one exists and is known.

Note that in some cases PRAGMA database_list is likely to be more robust.

pub fn execute_named( &self, sql: &str, params: &[(&str, &dyn ToSql)], ) -> Result<usize, Error>

👎Deprecated: You can use execute with named params now.

Convenience method to prepare and execute a single SQL statement with named parameter(s).

On success, returns the number of rows that were changed or inserted or deleted (via sqlite3_changes).

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn last_insert_rowid(&self) -> i64

Get the SQLite rowid of the most recent successful INSERT.

Uses sqlite3_last_insert_rowid under the hood.

pub fn query_row<T, P, F>(&self, sql: &str, params: P, f: F) -> Result<T, Error>
where P: Params, F: FnOnce(&Row<'_>) -> Result<T, Error>,

Convenience method to execute a query that is expected to return a single row.

Example

fn preferred_locale(conn: &Connection) -> Result<String> {
    conn.query_row(
        "SELECT value FROM preferences WHERE name='locale'",
        [],
        |row| row.get(0),
    )
}

If the query returns more than one row, all rows except the first are ignored.

Returns Err(QueryReturnedNoRows) if no results are returned. If the query truly is optional, you can call .optional() on the result of this to get a Result<Option<T>>.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn query_row_named<T, F>( &self, sql: &str, params: &[(&str, &dyn ToSql)], f: F, ) -> Result<T, Error>
where F: FnOnce(&Row<'_>) -> Result<T, Error>,

👎Deprecated: You can use query_row with named params now.

Convenience method to execute a query with named parameter(s) that is expected to return a single row.

If the query returns more than one row, all rows except the first are ignored.

Returns Err(QueryReturnedNoRows) if no results are returned. If the query truly is optional, you can call .optional() on the result of this to get a Result<Option<T>>.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn query_row_and_then<T, E, P, F>( &self, sql: &str, params: P, f: F, ) -> Result<T, E>
where P: Params, F: FnOnce(&Row<'_>) -> Result<T, E>, E: From<Error>,

Convenience method to execute a query that is expected to return a single row, and execute a mapping via f on that returned row with the possibility of failure. The Result type of f must implement std::convert::From<Error>.

Example

fn preferred_locale(conn: &Connection) -> Result<String> {
    conn.query_row_and_then(
        "SELECT value FROM preferences WHERE name='locale'",
        [],
        |row| row.get(0),
    )
}

If the query returns more than one row, all rows except the first are ignored.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn prepare(&self, sql: &str) -> Result<Statement<'_>, Error>

Prepare a SQL statement for execution.

Example

fn insert_new_people(conn: &Connection) -> Result<()> {
    let mut stmt = conn.prepare("INSERT INTO People (name) VALUES (?)")?;
    stmt.execute(["Joe Smith"])?;
    stmt.execute(["Bob Jones"])?;
    Ok(())
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn close(self) -> Result<(), (Connection, Error)>

Close the SQLite connection.

This is functionally equivalent to the Drop implementation for Connection except that on failure, it returns an error and the connection itself (presumably so closing can be attempted again).

Failure

Will return Err if the underlying SQLite call fails.

pub unsafe fn handle(&self) -> *mut sqlite3

Get access to the underlying SQLite database connection handle.

Warning

You should not need to use this function. If you do need to, please open an issue on the rusqlite repository and describe your use case.

Safety

This function is unsafe because it gives you raw access to the SQLite connection, and what you do with it could impact the safety of this Connection.

pub unsafe fn from_handle(db: *mut sqlite3) -> Result<Connection, Error>

Create a Connection from a raw handle.

The underlying SQLite database connection handle will not be closed when the returned connection is dropped/closed.

Safety

This function is unsafe because improper use may impact the Connection.

pub fn get_interrupt_handle(&self) -> InterruptHandle

Get access to a handle that can be used to interrupt long running queries from another thread.

pub fn is_autocommit(&self) -> bool

Test for auto-commit mode. Autocommit mode is on by default.

pub fn is_busy(&self) -> bool

Determine if all associated prepared statements have been reset.

pub fn cache_flush(&self) -> Result<(), Error>

Flush caches to disk mid-transaction