Module assembly_fdb::core::iter

Implementations of IntoIterator for the core model

Type Aliases

BucketRowIterMapper

A static pointer to a function from Bucket to a row iterator.

BucketRowRefIterMapper

A static pointer to a function from Bucket reference to a Row reference iterator.

FieldRefIter

An iterator over a slice of fields in a row reference.

FieldVecIter

An iterator over a vector of fields in a row.

RowSliceIter

An iterator over a slice of rows in a bucket reference.

RowVecIter

An iterator over a vector of rows in a bucket.

TableBucketIter

An iterator over a vector of buckets in a table.

TableBucketRefIter

An iterator over a slice of buckets in a table reference.

TableRowIter

A flattened iterator over all rows in a table, disregarding buckets.

TableRowRefIter

A flattened iterator over all row references in a table, disregarding buckets.