Module assembly_fdb::mem

Low-Level API that is suitable for non-little-endian machines

This is the default in-memory API the the FDB file format. It is useful for batch processing because it is fast and only loads the values that are accessed.

The reference structures in this module all implement Copy.

The only limitation is, that all references are bounded by the lifetime of the original database buffer.

Re-exports

pub use iter::FieldIter;

pub use iter::RowHeaderIter;

pub use iter::TableIter;

Modules

iter

Iterator types

Structs

Bucket

Reference to a single bucket

Column

Reference to a column definition

Database

A complete in-memory read-only database

Header

Reference to the tables array

MemContext

The context for mem::Field

Row

Reference to a single row

Table

Reference to a single table

Tables

Reference to the tables array

Type Definitions

Field

Value of or reference to a field value