Module assembly_fdb::mem::iter

Iterator types

This module contains the types that can be used to iterate over the structs in the assembly_fdb::mem module.

Structs

BucketIter

Iterator produced by Table::bucket_iter

FieldIter

An iterator over fields in a row

RowHeaderIter

Struct that implements Bucket::row_iter.

TableIter

Iterator created by Tables::iter

TableRowIter

Iterator produced by Table::row_iter