Module assembly_fdb::ro::buffer

General methods for aligned access to a byte buffer

Structs

Buffer

Wrapper around a immutable reference to a byte slice

Enums

BufferError

Error for handling a buffer

Functions

cmp_table_header_name

Compares the name given by bytes with the one referenced in table_header

get_at

Get a reference to a type at the given address of this buffer

get_slice_at

Get a reference to a slice at the given address of this buffer

header

Get the header of the file.

header_ref

Get the database header

table_data

Get the table data header

table_data_ref

Get the table data reference

table_definition

Get the table definition header

table_definition_ref

Get the table definition reference

table_headers

Get the table slice

Type Definitions

Res

Result with a BufferError