Struct assembly_fdb::ro::BaseHandle

pub struct BaseHandle<P: Deref, T>
where
    <P as Deref>::Target: AsRef<[u8]>,
{ /* private fields */ }

Base type for a handle to an in-memory data structure

This is basic layout of a handle to an in-memory FDB database. Internally, there is a pointer (&[u8]/Box<[u8]>/Rc<[u8]>/Arc<Mmap>/…) the memory slice as well as a value that represents the current target.

Implementations

impl<P: Deref, T> BaseHandle<P, T>
where P::Target: AsRef<[u8]>,

pub fn map_into<M, O, E>(self, map: M) -> BaseResult<P, O>
where M: Fn(&[u8], T) -> Result<O, E>, E: Into<BaseErrorKind>,

Get the tables

impl<P: Deref> BaseHandle<P, ()>
where P::Target: AsRef<[u8]>,

pub fn into_tables(self) -> BaseResult<P, FDBHeader>

Get the tables

impl<P: Deref> BaseHandle<P, FDBHeader>
where P::Target: AsRef<[u8]>,

pub fn into_table_at( self, index: usize, ) -> BaseResult<P, Option<FDBTableHeader>>

Get the tables

pub fn into_table_by_name( self, name: &Latin1Str, ) -> BaseResult<P, Option<FDBTableHeader>>

Get the tables

impl<P: Deref> BaseHandle<P, FDBTableHeader>
where P::Target: AsRef<[u8]>,

pub fn into_definition(self) -> BaseResult<P, FDBTableDefHeader>

Get the tables

pub fn into_data(self) -> BaseResult<P, FDBTableDataHeader>

Get the tables

impl<P: Deref> BaseHandle<P, FDBTableDataHeader>
where P::Target: AsRef<[u8]>,

pub fn get_bucket_for_hash(self, id: u32) -> BaseResult<P, FDBBucketHeader>

Get the bucket for a particular id / hash

impl<'a> BaseHandle<&'a [u8], ()>

pub fn new_ref(mem: &'a [u8]) -> Self

Create a new database handle

pub fn tables(&self) -> Result<'a, FDBHeader>

Get the header for the local database

impl<'a> BaseHandle<&'a [u8], FDBHeader>

pub fn table_count(&self) -> u32

Get the number of tables

pub fn table_header_list(&self) -> Result<'a, FDBTableHeaderSlice<'a>>

Get the table header slice

impl<'a> BaseHandle<&'a [u8], FDBTableHeader>

pub fn table_def_header(&self) -> Result<'a, FDBTableDefHeader>

Get the table definition header

pub fn table_data_header(&self) -> Result<'a, FDBTableDataHeader>

Get the table data header

impl<'a> BaseHandle<&'a [u8], FDBTableDefHeader>

pub fn column_count(&self) -> u32

Get the number of columns

pub fn table_name(&self) -> Result<'a, &'a Latin1Str>

Get the name of the table

pub fn column_header_list(&self) -> Result<'a, FDBColumnHeaderSlice<'a>>

Get the column header list

impl<'a> BaseHandle<&'a [u8], FDBColumnHeader>

pub fn column_name(&self) -> Result<'a, &'a Latin1Str>

Get the name of the column

pub fn column_data_type(&self) -> StdResult<ValueType, UnknownValueType>

Get the type of the column

impl<'a> BaseHandle<&'a [u8], FDBTableDataHeader>

pub fn bucket_count(&self) -> u32

Get the number of buckets

pub fn bucket_header_list(&self) -> Result<'a, FDBBucketHeaderSlice<'a>>

Get the slice of buckets

impl<'a> BaseHandle<&'a [u8], FDBBucketHeader>

pub fn first(&self) -> Option<Result<'a, FDBRowHeaderListEntry>>

Get the first row header entry or None

pub fn row_header_iter(&self) -> Handle<'a, FDBRowHeaderRef>

Get an iterator over all buckets

impl<'a> BaseHandle<&'a [u8], FDBRowHeaderListEntry>

pub fn next(&self) -> Option<Result<'a, FDBRowHeaderListEntry>>

Get the next row header list entry instance

pub fn row_header(&self) -> Result<'a, FDBRowHeader>

Get the associated row header.

impl<'a> BaseHandle<&'a [u8], FDBRowHeader>

pub fn field_count(&self) -> u32

Get the number of fields

pub fn field_data_list(&self) -> Result<'a, FDBFieldDataSlice<'a>>

Get the slice of fields

impl<'a> BaseHandle<&'a [u8], FDBFieldData>

pub fn try_get_value(&self) -> Result<'a, FDBFieldValue>

Get the value from this handle

impl<B: AsRef<[u8]>> BaseHandle<Arc<B>, ()>

pub fn new_arc(inner: B) -> Self

Create a new atomically-reference counted handle

impl<B: AsRef<[u8]>, T: Copy> BaseHandle<Arc<B>, T>

pub fn as_bytes_handle(&self) -> Handle<'_, T>

Borrow the atomically-reference counted handle as a byte handle

You can use this function to make cloning cheaper

impl<P: Deref> BaseHandle<P, ()>
where <P as Deref>::Target: AsRef<[u8]>,

pub fn new(mem: P) -> Self

Creates a new handle

impl<T, P: Deref> BaseHandle<P, Option<T>>
where <P as Deref>::Target: AsRef<[u8]>,

pub fn transpose(self) -> Option<BaseHandle<P, T>>

Turns a handle of an option into an option of a handle

impl<P: Deref, T> BaseHandle<P, T>
where <P as Deref>::Target: AsRef<[u8]>,

pub fn raw(&self) -> &T

Get a reference to the raw value inside

pub fn raw_mut(&mut self) -> &mut T

Get a reference to the raw value inside

pub fn as_bytes(&self) -> &[u8]

Get the byte slice for the whole database

pub fn replace<O>(self, raw: O) -> BaseHandle<P, O>

Replace the value that is stored with the memory pointer

impl<'a, T> BaseHandle<&'a [u8], T>

pub fn buf(self) -> &'a [u8]

Returns a copy of the contained buffer

pub fn into_raw(self) -> T

Get the raw value out of the handle

pub fn map<X>(self, mapper: impl Fn(&'a [u8], T) -> X) -> Handle<'a, X>

Map something with a closure

pub fn map_val<X>(self, mapper: impl Fn(T) -> X) -> Handle<'a, X>

Map the value with a closure

pub fn try_map<X, E>( self, mapper: impl Fn(&'a [u8], T) -> Result<X, E>, ) -> Result<Handle<'a, X>, E>

Map something with a closure

impl<'a, T> BaseHandle<&'a [u8], &'a [T]>

pub fn get(self, index: usize) -> Option<RefHandle<'a, T>>

Get the reference at index

impl<'a, T: Repr> BaseHandle<&'a [u8], &'a T>

pub fn map_extract(self) -> Handle<'a, T::Value>

Extract a value from a reference